+++ docs/changelog.md
@@ -10,6 +10,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - Useful features documented
+- `TypeVar`, `ParamSpec` and `TypeVarTuple` definitions are no longer added to `__all__`
 
 ## [0.4.0] - 2022-11-05